fix(search): spotlight and related-content follow-ups
- extract spotlight context + toast helpers into their own modules
- spotlight/toaster/content-marker polish; presentations page wiring
- related-content build: document the dependsOn ordering (apps must
  build first or the projects cache poisons deployment URLs)
- dedupe the missing-language-logo console log (once per language,
  it fires per card per render pass)
- update the search/related-content plan doc to match what landed
